+++ cvat/apps/engine/models.py
@@ -1,4 +1,5 @@
 # Copyright (C) 2018-2022 Intel Corporation
+# Copyright (C) 2022 CVAT.ai Corporation
 #
 # SPDX-License-Identifier: MIT
 
@@ -14,6 +15,8 @@ from django.core.files.storage import FileSystemStorage
 from django.db import models
 from django.db.models.fields import FloatField
 from django.core.serializers.json import DjangoJSONEncoder
+from drf_spectacular.types import OpenApiTypes
+from drf_spectacular.utils import extend_schema_field
 from cvat.apps.engine.utils import parse_specific_attributes
 from cvat.apps.organizations.models import Organization
 
@@ -37,6 +40,8 @@ class DimensionType(str, Enum):
         return self.value
 
 class StatusChoice(str, Enum):
+    """Deprecated. Use StageChoice and StateChoice instead"""
+
     ANNOTATION = 'annotation'
     VALIDATION = 'validation'
     COMPLETED = 'completed'
@@ -435,6 +440,7 @@ class Job(models.Model):
     def get_dirname(self):
         return os.path.join(settings.JOBS_ROOT, str(self.id))
 
+    @extend_schema_field(OpenApiTypes.INT)
     def get_project_id(self):
         project = self.segment.task.project
         return project.id if project else None